# mindbridge/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MindBridge API",
    "description": "RESTful API for the MindBridge AI-guided CBT platform for workplace burnout recovery. Covers authentication, onboarding, CBT sessions, thought record diary, progress tracking, crisis support, media analysis, personalization, and healthcare integration.",
    "version": "1.0.0",
    "contact": {
      "name": "MindBridge Engineering",
      "email": "api@mindbridge.health"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.mindbridge.health/v1",
      "description": "Production"
    },
    {
      "url": "https://staging-api.mindbridge.health/v1",
      "description": "Staging"
    }
  ],
...
  "paths": {
...
    "/diary/entries": {
      "get": {
        "tags": ["Diary"],
        "operationId": "getDiaryEntries",
        "summary": "List all diary entries",
        "description": "Returns all thought record diary entries with date, mood shift, emotions, and distortion tags. Supports US-3.1 and REQ-DRY-001.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": { "type": "string", "enum": ["date_desc", "date_asc", "mood_change"], "default": "date_desc" }
          }
        ],
        "responses": {
          "200": {
            "description": "Diary entries retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DiaryEntrySummary" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Diary"],
        "operationId": "createDiaryEntry",
        "summary": "Create a new diary entry",
        "description": "Submits the complete diary entry (mood, situation, thoughts, emotions, distortions, reframe, belief rating). Supports US-3.2 through US-3.7 and REQ-DRY-002 through DRY-007.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diary entry created with AI insights",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryEntryCreateResponse"
                }
              }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/diary/entries/{entryId}": {
      "get": {
        "tags": ["Diary"],
        "operationId": "getDiaryEntry",
        "summary": "Get diary entry detail with AI analysis",
        "description": "Returns the full diary entry with overview, AI-guided transcript, and AI analysis tabs. Supports US-3.9 and REQ-DRY-009.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Diary entry detail retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryEntryDetail"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Diary"],
        "operationId": "deleteDiaryEntry",
        "summary": "Delete a diary entry",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Entry deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/diary/distortions/suggest": {
      "post": {
        "tags": ["Diary"],
        "operationId": "suggestDistortions",
        "summary": "AI-suggest likely distortions from a thought",
        "description": "Analyzes an automatic thought and returns suggested cognitive distortions. Supports US-3.5 and REQ-DRY-005.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["thought"],
                "properties": {
                  "thought": { "type": "string", "maxLength": 2000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Suggested distortions returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "distortionId": { "type": "string" },
                          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
                          "reasoning": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/diary/insights": {
      "get": {
        "tags": ["Diary"],
        "operationId": "getDiaryInsights",
        "summary": "Get pattern insights across all diary entries",
        "description": "Returns aggregated mood trends, top emotions, top distortions, and AI narrative analysis. Supports US-3.8 and REQ-DRY-008.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Pattern insights retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryInsights"
                }
              }
            }
          }
        }
      }
    },
...
    "/crisis": {
      "get": {
        "tags": ["Crisis"],
        "operationId": "getCrisisHub",
        "summary": "Get crisis support hub data",
        "description": "Returns crisis awareness content, pathways, and 988 call action. Supports US-5.1 and REQ-CRS-001/002.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Crisis hub data retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrisisHub"
                }
              }
            }
          }
        }
      }
    },
...
    "/crisis/safety-plan": {
      "get": {
        "tags": ["Crisis"],
        "operationId": "getSafetyPlan",
        "summary": "Get personalized safety plan",
        "description": "Returns the 6-step safety plan with warning signs, coping strategies, contacts, resources, and environment safety. Supports US-5.4 and REQ-CRS-008/009.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Safety plan retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyPlan"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Crisis"],
        "operationId": "updateSafetyPlan",
        "summary": "Update the personalized safety plan",
        "description": "Updates one or more steps of the safety plan. Supports US-5.4.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyPlanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Safety plan updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyPlan"
                }
              }
            }
          }
        }
      }
    },
    "/crisis/detect": {
      "post": {
        "tags": ["Crisis"],
        "operationId": "detectCrisis",
        "summary": "Analyze text for crisis indicators",
        "description": "Scans user input for crisis keywords and sentiment patterns. Returns risk level and recommended action. Supports REQ-CRS-010.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": { "type": "string" },
                  "sessionId": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Crisis analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrisisDetectionResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": ["code", "message", "errors"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
      "DiaryEntrySummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "moodBefore": { "type": "integer", "minimum": 1, "maximum": 10 },
          "moodAfter": { "type": "integer", "minimum": 1, "maximum": 10 },
          "emotions": { "type": "array", "items": { "type": "string" } },
          "distortions": { "type": "array", "items": { "type": "string" } },
          "situationPreview": { "type": "string", "description": "Truncated first 100 chars of situation" }
        }
      },
      "DiaryEntryCreate": {
        "type": "object",
        "required": ["moodBefore", "situation", "automaticThought", "emotions", "distortionIds", "balancedThought", "beliefRating", "moodAfter"],
        "properties": {
          "moodBefore": { "type": "integer", "minimum": 1, "maximum": 10 },
          "situation": {
            "type": "object",
            "required": ["description"],
            "properties": {
              "description": { "type": "string", "maxLength": 5000 },
              "timeOfDay": { "type": "string" },
              "location": { "type": "string" }
            }
          },
          "automaticThought": { "type": "string", "maxLength": 2000 },
          "emotions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": ["name", "intensity"],
              "properties": {
                "name": { "type": "string" },
                "intensity": { "type": "integer", "minimum": 1, "maximum": 10 }
              }
            }
          },
          "distortionIds": { "type": "array", "minItems": 1, "items": { "type": "string" } },
          "balancedThought": { "type": "string", "maxLength": 2000 },
          "beliefRating": { "type": "integer", "minimum": 0, "maximum": 100 },
          "moodAfter": { "type": "integer", "minimum": 1, "maximum": 10 }
        }
      },
      "DiaryEntryCreateResponse": {
        "type": "object",
        "properties": {
          "entry": { "$ref": "#/components/schemas/DiaryEntrySummary" },
          "moodChange": {
            "type": "object",
            "properties": {
              "before": { "type": "integer" },
              "after": { "type": "integer" },
              "direction": { "type": "string", "enum": ["improved", "declined", "unchanged"] }
            }
          },
          "aiInsights": {
            "type": "object",
            "properties": {
              "primaryDistortion": { "type": "string" },
              "triggerPattern": { "type": "string" },
              "encouragement": { "type": "string" }
            }
          }
        }
      },
      "DiaryEntryDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "moodBefore": { "type": "integer" },
          "moodAfter": { "type": "integer" },
          "situation": {
            "type": "object",
            "properties": {
              "description": { "type": "string" },
              "timeOfDay": { "type": "string" },
              "location": { "type": "string" }
            }
          },
          "automaticThought": { "type": "string" },
          "emotions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "intensity": { "type": "integer" }
              }
            }
          },
          "distortions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CognitiveDistortion" }
          },
          "balancedThought": { "type": "string" },
          "beliefRating": { "type": "integer" },
          "aiAnalysis": {
            "type": "object",
            "properties": {
              "summary": { "type": "string" },
              "insights": { "type": "array", "items": { "type": "string" } },
              "recommendations": { "type": "array", "items": { "type": "string" } }
            }
          }
        }
      },
      "CognitiveDistortion": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "examples": { "type": "array", "items": { "type": "string" } }
        }
      },
      "DiaryInsights": {
        "type": "object",
        "properties": {
          "totalEntries": { "type": "integer" },
          "sufficientData": { "type": "boolean" },
          "moodTrend": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string", "format": "date" },
                "averageMood": { "type": "number" }
              }
            }
          },
          "topEmotions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "count": { "type": "integer" },
                "percentage": { "type": "number" }
              }
            }
          },
          "topDistortions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "count": { "type": "integer" },
                "percentage": { "type": "number" }
              }
            }
          },
          "aiNarrative": { "type": "string", "description": "AI-generated pattern analysis narrative" }
        }
      },
...
      "CrisisHub": {
        "type": "object",
        "properties": {
          "awarenessCard": {
            "type": "object",
            "properties": {
              "headline": { "type": "string" },
              "supportMessage": { "type": "string" },
              "emergencyNumber": { "type": "string" }
            }
          },
          "pathways": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "enum": ["coping_strategies", "talk_to_someone", "safety_plan"] },
                "label": { "type": "string" },
                "description": { "type": "string" },
                "icon": { "type": "string" }
              }
            }
          },
          "crisisLineAction": {
            "type": "object",
            "properties": {
              "label": { "type": "string" },
              "phoneNumber": { "type": "string" },
              "available24x7": { "type": "boolean" }
            }
          }
        }
      },
...
      "SafetyPlan": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "stepNumber": { "type": "integer", "minimum": 1, "maximum": 6 },
                "title": { "type": "string" },
                "items": { "type": "array", "items": { "type": "string" } }
              }
            }
          },
          "reasonForLiving": { "type": "string" },
          "lastUpdated": { "type": "string", "format": "date-time" }
        }
      },
      "SafetyPlanUpdate": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "stepNumber": { "type": "integer", "minimum": 1, "maximum": 6 },
                "items": { "type": "array", "items": { "type": "string" } }
              }
            }
          },
          "reasonForLiving": { "type": "string" }
        }
      },
      "CrisisDetectionResult": {
        "type": "object",
        "properties": {
          "riskLevel": { "type": "string", "enum": ["none", "low", "medium", "high", "critical"] },
          "keywordsDetected": { "type": "array", "items": { "type": "string" } },
          "recommendedAction": { "type": "string", "enum": ["none", "show_resources", "show_crisis_hub", "immediate_intervention"] },
          "message": { "type": "string", "nullable": true }
        }
      },
...
    }
  }
}