# src/main/java/com/digitaltherapy/dto/DiaryEntryDetail.java
package com.digitaltherapy.dto;

import java.time.LocalDateTime;
import java.util.List;
import java.util.UUID;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class DiaryEntryDetail {

    private UUID id;
    private String situation;
    private String automaticThought;
    private List<EmotionInfo> emotions;
    private List<DistortionInfo> distortions;
    private String alternativeThought;
    private Integer moodBefore;
    private Integer moodAfter;
    private Integer beliefRatingBefore;
    private Integer beliefRatingAfter;
    private LocalDateTime createdAt;

    @Data
    @Builder
    @NoArgsConstructor
    @AllArgsConstructor
    public static class EmotionInfo {

        private String emotion;
        private Integer intensity;
    }

    @Data
    @Builder
    @NoArgsConstructor
    @AllArgsConstructor
    public static class DistortionInfo {

        private String id;
        private String name;
        private String description;
    }
}